Sort out duplex operations
LoRa.receive() needs to be called after sending a packet,
to reenable interrupts.

The test programs were combined into a single one which
will transmit at rate specified by a compile define.

However, when running a pair of duplex programs (one
faster than the other), the one that is transmitting faster
drops incoming messages occaisionally. I suspect that
it's transmitted message is stepping on the incoming
message.

Perhaps it will be more reliable if one side of the pair
is sending very short messages.